rag-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-community==0.0.10
psycopg2-binary==2.9.9
pgvector==0.2.4
redis==5.0.1
celery==5.3.4
pika==1.3.2
sentence-transformers==2.2.2
torch==2.1.1
transformers==4.36.2
llama-cpp-python==0.2.20
unsloth
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2
neo4j==5.15.0
